flaskBackendDatabase/requirements.txt
pyodbc
dotenv
flask
flask_restful
tenacity
opencensus-ext-azure
opencensus-ext-flask
logging
azure-identity
werkzeug